Conflicts:
	Dockerfile
	config/server.properties.default
	start.sh

+++ start.sh
@@ -1,15 +1,49 @@
-#!/bin/bash
+#!/bin/bash -x
 
 # Necessary?
 export JMX_PORT=7203
 
-echo "ZOOKEEPER_IP=$ZOOKEEPER_PORT_2181_TCP_ADDR"
+EXTENSION=""
+case $BRANCH in
+  master)
+    EXTENSION=".prod"
+    CHROOT="/v0_8_1"
 
-echo "Fixing ZOOKEEPER_IP in server.properties"
-cat /kafka/config/server.properties.tmpl \
-  | sed "s|{{ZOOKEEPER_IP}}|${ZOOKEEPER_PORT_2181_TCP_ADDR:-localhost}|g" \
-  | sed "s|{{EXPOSED_HOST}}|${EXPOSED_HOST:-localhost}|g" \
-  > /kafka/config/server.properties
+    # TODO Service discovery
+    ZOOKEEPER_PORT=${ZOOKEEPER_PORT:-2181}
+  ;;
+  staging)
+    EXTENSION=".staging"
+    CHROOT="/v0_8_1"
+
+    # TODO Service discovery
+    ZOOKEEPER_PORT=${ZOOKEEPER_PORT:-2181}
+  ;;
+  *)
+    # Developer environments, etc.
+    EXTENSION=".default"
+    ZOOKEEPER_IP=${ZOOKEEPER_IP:-$ZOOKEEPER_PORT_2181_TCP_ADDR}
+    ZOOKEEPER_PORT=${ZOOKEEPER_PORT:-$ZOOKEEPER_PORT_2181_TCP_PORT}
+
+  ;;
+esac
+
+if [ -z "$ZOOKEEPER_IP" -o -z "$ZOOKEEPER_PORT" ]; then
+  echo "No valid ZOOKEEPER IP and port set! ${BRANCH}"
+  exit -1
+fi
+
+
+IP=$(cat /etc/hosts | head -n1 | awk '{print $1}')
+PORT=9092
+
+cat /kafka/config/server.properties${EXTENSION} \
+  | sed "s|{{ZOOKEEPER_IP}}|${ZOOKEEPER_IP}|g" \
+  | sed "s|{{BROKER_ID}}|${BROKER_ID:-0}|g" \
+  | sed "s|{{CHROOT}}|${CHROOT:-}|g" \
+  | sed "s|{{EXPOSED_HOST}}|${EXPOSED_HOST:-$IP}|g" \
+  | sed "s|{{EXPOSED_PORT}}|${EXPOSED_PORT:-9092}|g" \
+   > /kafka/config/server.properties
 
 echo "Starting kafka"
 /kafka/bin/kafka-server-start.sh /kafka/config/server.properties 2>&1 | tee /logs/kafka.log
